cmake-format: include necessary tooling in config Py
Makes this easier to ship with OOT modules
